fix(web): satisfy storybook page helper lint guard
